// packages/db/src/schema.d.ts
export declare const schema: {
    logLevelEnum: import("drizzle-orm/pg-core").PgEnum<["trace", "debug", "info", "warn", "error", "fatal"]>;
    httpMethodEnum: import("drizzle-orm/pg-core").PgEnum<["GET", "POST", "PUT", "DELETE", "PATCH", "HEAD", "OPTIONS"]>;
    systemLogs: import("drizzle-orm/pg-core").PgTableWithColumns<{
        name: "system_logs";
        schema: undefined;
        columns: {
// ...
            userId: import("drizzle-orm/pg-core").PgColumn<{
                name: "user_id";
                tableName: "system_logs";
                dataType: "string";
                columnType: "PgText";
                data: string;
                driverParam: string;
                notNull: false;
                hasDefault: false;
                isPrimaryKey: false;
                isAutoincrement: false;
                hasRuntimeDefault: false;
                enumValues: [string, ...string[]];
                baseColumn: never;
                generated: undefined;
            }, {}, {}>;
// ...
        };
        dialect: "pg";
    }>;
// ...
};